+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/data/external/CardsBroker.kt
@@ -1,4 +1,4 @@
-package ayds.lisboa.songinfo.moredetails.data
+package ayds.lisboa.songinfo.moredetails.data.external
 
 import ayds.lisboa.songinfo.moredetails.data.proxy.ProxyCard
 import ayds.lisboa.songinfo.moredetails.domain.entities.Card
